+++ app/src/main/java/com/example/object_oriented_pj_10/EXERCISE/MyExerciseRepository.kt
@@ -1,0 +1,5 @@
+package com.example.object_oriented_pj_10.EXERCISE
+
+class MyExerciseRepository {
+
+}
