+++ app/src/main/java/com/example/object_oriented_pj_10/MEMO/MemoFragment.kt
@@ -1,21 +1,16 @@
-package com.example.object_oriented_pj_10
+package com.example.object_oriented_pj_10.MEMO
 
 import android.os.Build
 import android.os.Bundle
-import android.os.Handler
-import android.util.Log
 import android.view.LayoutInflater
 import android.view.View
 import android.view.ViewGroup
 import androidx.annotation.RequiresApi
 import androidx.fragment.app.Fragment
-import androidx.fragment.app.activityViewModels
 import androidx.fragment.app.viewModels
 import androidx.recyclerview.widget.LinearLayoutManager
 import com.example.object_oriented_pj_10.databinding.FragmentMemoBinding
-import com.example.object_oriented_pj_10.repository.MyMemoRepository
 import kotlinx.android.synthetic.main.fragment_memo.*
-import kotlinx.coroutines.delay
 import java.time.LocalDate
 import java.util.*
