+++ app/src/main/java/com/example/object_oriented_pj_10/MEMO/MemoList.kt
@@ -1,4 +1,4 @@
-package com.example.object_oriented_pj_10
+package com.example.object_oriented_pj_10.MEMO
 
 data class MemoList (
     val title : String,
