--- app/src/main/java/com/example/object_oriented_pj_10/repository/MyExerciseRepository.kt
@@ -1,9 +1,0 @@
-package com.example.object_oriented_pj_10.repository
-
-import com.example.object_oriented_pj_10.ExerciseList
-import com.google.firebase.database.ktx.database
-import com.google.firebase.ktx.Firebase
-
-class MyExerciseRepository {
-
-}
