+++ app/src/main/res/layout/fragment_cook.xml
@@ -4,7 +4,7 @@
     xmlns:tools="http://schemas.android.com/tools"
     android:layout_width="match_parent"
     android:layout_height="match_parent"
-    tools:context=".CookFragment">
+    tools:context=".COOK.CookFragment">
 
     <LinearLayout
         android:id="@+id/linearLayout7"
